requirements.txt
openai
streamlit
pandas 
langchain
Pillow
